+++ examples/storybook/stories/Components/EditorWrapper.js
@@ -151,17 +151,17 @@ const EditorWrapper = ({
   const editorPlugins = pluginsToDisplay
     ? pluginsToDisplay.map(plugin => pluginsMap[plugin])
     : plugins;
-  const theme = palette ? { theme: 'Palette', palette } : { theme: 'Default' };
   return (
     <RicosEditor
       plugins={editorPlugins}
-      {...theme}
+      theme={{ palette }}
       content={content}
       isMobile={isMobile}
       placeholder={'Share something...'}
       toolbarSettings={toolbarSettings}
+      onChange={onChange}
     >
-      <RichContentEditor helpers={{ onFilesChange }} onChange={onChange} config={config} />
+      <RichContentEditor helpers={{ onFilesChange }} config={config} />
     </RicosEditor>
   );
 };