+++ packages/ricos-editor/web/src/RicosEditor.tsx
@@ -16,9 +16,9 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
   editor: typeof RichContentEditor;
   dataInstance: EditorDataInstance;
 
-  constructor(props) {
+  constructor(props: RicosEditorProps) {
     super(props);
-    this.dataInstance = createDataConverter();
+    this.dataInstance = createDataConverter(props.onChange);
     this.state = {};
   }
 
@@ -36,14 +36,12 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
 
   focus = () => this.editor.focus();
   blur = () => this.editor.blur();
-  getData = (postId?: string, forPublish?: boolean) => {
+  getContent = (postId?: string, forPublish?: boolean) => {
     const { getContentState } = this.dataInstance;
     if (postId && forPublish) {
       this.editor.publish(postId); //async
     }
-    return {
-      getContentState,
-    };
+    return getContentState();
   };
 
   render() {