+++ CHANGELOG.md
@@ -26,8 +26,10 @@
   - [#1084](https://github.com/wix-incubator/rich-content/pull/1084) external toolbar API added
 - `wrapper`
   - [#1003](https://github.com/wix-incubator/rich-content/pull/1003) RichContentWrapper was split into RicosEditor & RicosViewer
+  - [#1115](https://github.com/wix-incubator/rich-content/pull/1115) Improved RicosEditor API of getContent + onChange
 - `viewer`
-  - [#1093](https://github.com/wix-incubator/rich-content/pull/1093) viewerAction callback in helpers is now supported, triggered on image expand (gallery & viewer)  
+  - [#1088](https://github.com/wix-incubator/rich-content/pull/1088) viewerAction callback in helpers is now supported, triggered on image expand (gallery & viewer)
+  - [#1093](https://github.com/wix-incubator/rich-content/pull/1093) viewerAction callback in helpers is now supported, triggered on image expand (gallery & viewer)
   - [#1116](https://github.com/wix-incubator/rich-content/pull/1116) renamed onViewerAction to onAction + order of arguments
 ### :bug: Bug Fix
 - `wrapper` `viewer`
@@ -35,11 +37,15 @@
 - `plugins`
   - [#1084](https://github.com/wix-incubator/rich-content/pull/1084) svg icon id conflicts fixed
 - `giphy`
-  - [#1084](https://github.com/wix-incubator/rich-content/pull/1084) giphy toolbar not showing after giphy is added 
+  - [#1084](https://github.com/wix-incubator/rich-content/pull/1084) giphy toolbar not showing after giphy is added
 - `link-preview`
   - [#1048](https://github.com/wix-incubator/rich-content/pull/1048) fix: enable link preview and link embed by default
+- `file-upload`
+  - [#1082](https://github.com/wix-incubator/rich-content/pull/1082) error handling fix
 - `common`
   - [#1092](https://github.com/wix-incubator/rich-content/pull/1092) text indentation not showing in rtl
+- `html plugin`
+  - [#1126](https://github.com/wix-incubator/rich-content/pull/1126) fix html iframe src hight and remove html initial state
 ### :house: Internal
 - `exampleApp`
   - [0f849222](https://github.com/wix-incubator/rich-content/commit/0f849222) fix translations